Add info for MacPorts package manager
I think there should be a sub-page install-mac.rst that describes some details for MacPorts, Homebrew and Fink.

E.g. for Macports it's good to know that the Macports installer adds this to .profile:
export PATH=/opt/local/bin:/opt/local/sbin:$PATH

Also users will probably want to run these commands for the packages that install command line tools to get symlinks with the usual names (i.e. without version numbers in the names):
    sudo port select --set python python27
    sudo port select --set ipython ipython27
    sudo port select py-sympy py27-sympy
    sudo port select nosetests nosetests27

And very soon when maplotlib 1.2 is out the scipy stack will be available for 3.2 if the user wants, so that should be described.

I'm not a Fink and Homebrew user, but I can try it out and write detailed instructions for those also if you think it's useful.