DOC: Update library links + FAQ in documentation
Many links are now dead or need updating. This commit
does that for a lot of them in the topical software page.

This commit also patches a mistake in the astropy fits
package naming in the FAQ as well as adds the BSD license
link and the ufunc 'at' to a question regarding incrementing
an array by a scalar.